+++ example/lib/main.dart
@@ -22,6 +22,6 @@ class _App extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp.router(routerDelegate: Screen.delegate);
+    return MaterialApp.router(routerDelegate: Screen.manager);
   }
 }